requirements.txt
pyinstaller>=5.13.0
devicons==0.1.0
simple_term_menu==1.6.1
PyYAML==6.0.1
